requirements.txt
pandas==1.3.2
numpy==1.21.2
google-api-python-client==2.17.0
oauth2client==4.1.3
streamlit==0.87.0
